King's Tropism to Passed Pawns (#83)
Bench: 7488022

ELO | 5.97 +- 4.38 (95%)
SPRT | 10.0+0.10s Threads=1 Hash=16MB
LLR | 2.98 (-2.94, 2.94) [0.00, 5.00]
Games | N: 11120 W: 2654 L: 2463 D: 6003

A common guider to a pawns promotion is the king. For this reason it is important to look at the distances of the kings from a passed pawn (specifically it's next square). The closer your king is, the better, the closer the enemy king, the worse. The CPW link below discusses Manhattan distance. It is not implemented this way in Koivisto just controlling the advance square is helpful enough already.